+++ src/app/connections/page.tsx
@@ -1,7 +1,7 @@
 import HubTemplate from "@/templates/HubTemplate";
 import { getContent } from "@/utils/functions";
 import PageHeader from "@/components/PageHeader/PageHeader";
-import FullProjectCard from "@/components/ProjectCard/FullProjectCard";
+import ProjectBigCard from "@/components/ProjectBigtCard/ProjectBigCard";
 import { Project } from "@/utils/interfaces";
 
 export const revalidate = 60;
@@ -21,7 +21,7 @@ export default async function Connections({}: {}) {
         )}
       />
       {partners?.map((partner: { fields: Project }, index: number) => (
-        <FullProjectCard
+        <ProjectBigCard
           key={partner.fields.name}
           project={partner}
           direction={index % 2 === 0 ? "left" : "right"}
